+++ changelog.md
@@ -1,3 +1,15 @@
+# 1.21.4-0.10.0.0
+* Added Mechanical Graph API which works similarly to the signal graph
+* Mechanical Components can be assigned via the exmachina:block/mechanical_component datamap
+* Mechanical Component Types can be registered to the exmachina:mechanical_component_type registry, which define the nodes provided by a blockstate, their torque/inertia, and what connections they can form
+* Mechanical graph updates can be scheduled via ExMachinaGameEvents.scheduleMechanicalGraphUpdate or by triggering the exmachina:mechanical_graph_update game event
+* Added max_mechanical_graph_size to common config
+* Added exmachina:none mechanical component type (all blocks without components use this as a fallback)
+* Moved the helper method for scheduling signal graph updates to ExMachinaGameEvents
+
+# 1.21.4-0.9.0.0
+* Update to 1.21.4
+
 # 1.21.3-0.8.0.2
 * Fix origin nodes' source power being ignored in signal graphs